moved common certs code to encrypter
